backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
pymupdf
openai
boto3
chromadb
networkx
scikit-learn
numpy
python-dotenv
aiofiles
httpx
playwright
python-pptx
markdown
duckduckgo-search
ddgs
google-genai
